Clean up and split classes into separate files
